wip
category create almost done, still need work on slug logic

+++ internal/adapters/input/http/dto/categoryResponse.go
@@ -1,7 +1,7 @@
 package dto
 
 type CategoryResponse struct {
-	Id        int32  `json:"id"`
+	Id        int64  `json:"id"`
 	Name      string `json:"name"`
 	Slug      string `json:"slug"`
 	CreatedAt string `json:"created_at"`
